apps/bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.2
google-genai==1.14.0
redis[hiredis]==5.1.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
